check levelstat txt exists before attempting to remove it

+++ jacodemon/stats.py
@@ -119,7 +119,8 @@ class Statistics:
 
 def NewStatistics(launch: LaunchConfig, demo_dir: str) -> Statistics:
     
-    os.remove(LEVELSTAT_TXT)
+    if os.path.exists(LEVELSTAT_TXT):
+        os.remove(LEVELSTAT_TXT)
     statistics = Statistics(launch.timestamp, launch.get_comp_level(), 
                             launch.get_port(), launch.get_command(), 
                             launch.get_demo_name(), demo_dir)
